pyproject.toml
[project]
name = "voice-synced-text-display"
version = "0.1.0"
description = "Real-time voice-synced translated text display system"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "openai-whisper==20231117",
    "sounddevice==0.4.6",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "jinja2==3.1.2",
    "python-dotenv==1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "data*", "templates*"]

[tool.uv]
dev-dependencies = [] 